+++ RELEASE_NOTES.md
@@ -10,7 +10,8 @@ Update of the Pagination `Params` data class.
 
 ## New Features
 
-* Additional information for energy metric
+* Additional information for energy metric.
+* Add component state codes.
 
 ## Bug Fixes